🚚 refactor: TimestampLabel을 shared/ui/timestamp로 이동
